+++ src/Kiota.Builder/Writers/CodeTypeWriter.cs
@@ -1,3 +1,4 @@
+using Kiota.Builder.Writers.CSharp;
 using Kiota.Builder.Writers.TypeScript;
 
 namespace Kiota.Builder.Writers {
@@ -8,6 +9,8 @@ namespace Kiota.Builder.Writers {
         {
             if(conventions is TypeScriptConventionService tsConventions)
                 writer.Write(tsConventions.GetTypeString(codeElement, codeElement), includeIndent: false);
+            else if (conventions is CSharpConventionService csConventions)
+                writer.Write(csConventions.GetTypeString(codeElement, codeElement), includeIndent: false);
             else
                 writer.Write(conventions.GetTypeString(codeElement), includeIndent: false);
         }